test/Practice/model/requirements-dev.txt
# Fire Simulation Environment Configuration

# 개발 환경
-e .

# 과학 계산
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 시각화
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 지리공간 데이터
geopandas>=0.9.0
rasterio>=1.2.0
shapely>=1.7.0
pyproj>=3.0.0
folium>=0.12.0

# 기계학습
scikit-learn>=1.0.0

# 이미지 처리
Pillow>=8.0.0
opencv-python>=4.5.0

# 애니메이션
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# 성능 최적화
numba>=0.55.0
cython>=0.29.0

# 병렬처리
joblib>=1.0.0
multiprocessing-logging>=0.3.0

# 설정 관리
pyyaml>=5.4.0
python-dotenv>=0.19.0
configargparse>=1.5.0

# 로깅
loguru>=0.5.0

# 프로그레스 바
tqdm>=4.62.0

# 메모리 프로파일링
memory-profiler>=0.60.0
psutil>=5.8.0

# 웹 인터페이스 (선택사항)
streamlit>=1.10.0
dash>=2.0.0

# Jupyter 지원
jupyter>=1.0.0
ipywidgets>=7.6.0
